+++ code/packages/python/spice-engine/src/spice_engine/model_cards.py
@@ -307,8 +307,8 @@ _MODEL_CARD_SUPPORTED_PARAMETER_KINDS = (
 )
 _MODEL_CARD_SUPPORTED_PARAMETER_COVERAGE_EXPECTED_SUMMARIES = {
     "D": (12, 18, 5, 3),
-    "NPN": (21, 36, 11, 4),
-    "PNP": (21, 36, 11, 4),
+    "NPN": (23, 38, 11, 4),
+    "PNP": (23, 38, 11, 4),
     "NJF": (5, 11, 5, 3),
     "PJF": (5, 11, 5, 3),
     "NMOS": (18, 25, 6, 3),
@@ -380,6 +380,8 @@ _BJT_PARAMETER_ALIASES: dict[str, str] = {
     "IK": "IKF",
     "ISE": "ISE",
     "NE": "NE",
+    "ISC": "ISC",
+    "NC": "NC",
     "NF": "NF",
     "NR": "NR",
     "VJE": "VJE",
@@ -929,6 +931,8 @@ def bjt_from_model_card(
         Ikf=p.get("IKF", 0.0),
         Ise=p.get("ISE", 0.0),
         Ne=p.get("NE", 1.0),
+        Isc=p.get("ISC", 0.0),
+        Nc=p.get("NC", 2.0),
     )